+++ src/services/UserService.js
@@ -1,13 +1,29 @@
 import prisma from "../repositories/prisma.js";
 import { hashPassword } from "../auth/password.js";
 import { userToDTO } from "../dtos/userDTO.js";
+import { ConflictError, NotFoundError } from "../utils/customErrors.js";
 
 class UserService {
+  async getAllUsers() {
+    const users = await prisma.user.findMany();
+    return users.map(userToDTO);
+  }
+
+  async findUserById(id) {
+    const user = await prisma.user.findUnique({ where: { id } });
+    return user ? userToDTO(user) : null;
+  }
+
+  async findUserByNickname(nickname) {
+    const user = await prisma.user.findUnique({
+      where: { nickname },
+    });
+    return user ? userToDTO(user) : null;
+  }
+
   async createUser(nickname, password) {
     const existingUser = await prisma.user.findUnique({ where: { nickname } });
-    if (existingUser) {
-      throw new Error("Nickname já está em uso.");
-    }
+    if (existingUser) throw new ConflictError("Nickname já está em uso.");
 
     const hashedPassword = await hashPassword(password);
 
@@ -21,18 +37,6 @@ class UserService {
     return userToDTO(user);
   }
 
-  async findUserByNickname(nickname) {
-    const user = await prisma.user.findUnique({
-      where: { nickname },
-    });
-    return user ? userToDTO(user) : null;
-  }
-
-  async getAllUsers() {
-    const users = await prisma.user.findMany();
-    return users.map(userToDTO);
-  }
-
   async updateUser(id, data) {
     const user = prisma.user.update({
       where: { id },
@@ -42,6 +46,9 @@ class UserService {
   }
 
   async deleteUser(id) {
+    const user = await this.findUserById(id);
+    if (!user) throw new NotFoundError("Usuário não encontrado.");
+
     return prisma.user.delete({
       where: { id },
     });
